backend/requirements.txt
Django==4.2
django-colorfield==0.8.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==6.0.0
requests==2.28.2
webcolors==1.13
